# project2/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Users & Products API",
    "description": "API Documentation for managing users and products.",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
    "/users/": {
      "get": {
        "description": "",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": { "example": "any" },
                "lastName": { "example": "any" },
                "email": { "example": "any" },
                "favoriteColor": { "example": "any" },
                "birthday": { "example": "any" }
              }
            }
          }
        ],
        "responses": {
          "201": { "description": "Created" },
          "412": { "description": "Precondition Failed" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "description": "",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Bad Request" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": { "example": "any" },
                "lastName": { "example": "any" },
                "email": { "example": "any" },
                "favoriteColor": { "example": "any" },
                "birthday": { "example": "any" }
              }
            }
          }
        ],
        "responses": {
          "204": { "description": "No Content" },
          "400": { "description": "Bad Request" },
          "404": { "description": "Not Found" },
          "412": { "description": "Precondition Failed" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": { "description": "No Content" },
          "400": { "description": "Bad Request" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/products/": {
      "get": {
        "description": "Get all products",
        "produces": ["application/json"],
        "responses": {
          "200": { "description": "A list of products" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "post": {
        "description": "Create a new product",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["name", "price"],
              "properties": {
                "name": { "type": "string", "example": "Laptop" },
                "price": { "type": "number", "example": 799.99 },
                "description": { "type": "string", "example": "High-end device" }
              }
            }
          }
        ],
        "responses": {
          "201": { "description": "Product created successfully" },
          "400": { "description": "Invalid input" }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "description": "Get product by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": { "description": "Product found" },
          "404": { "description": "Product not found" }
        }
      },
      "put": {
        "description": "Update a product",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["name", "price"],
              "properties": {
                "name": { "type": "string", "example": "Updated Laptop" },
                "price": { "type": "number", "example": 699.99 },
                "description": { "type": "string", "example": "Updated description" }
              }
            }
          }
        ],
        "responses": {
          "204": { "description": "Product updated successfully" },
          "400": { "description": "Invalid product data" },
          "404": { "description": "Product not found" }
        }
      },
      "delete": {
        "description": "Delete a product",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": { "description": "Product deleted" },
          "404": { "description": "Product not found" }
        }
      }
    }
  }
}